+++ .gitignore
@@ -1,5 +1,5 @@
 __pycache__/
-mini_bdx_runtime.egg-info/
+open_duck_mini_runtime.egg-info/
 build
 dist
 .DS_Store
