+++ setup.cfg
@@ -1,10 +1,10 @@
 [metadata]
-name = mini-bdx-runtime
-version = 0.0.1
+name = open-duck-mini-runtime
+version = 1.0.0
 author = Antoine Pirrone
 author_email = antoine.pirrone@gmail.com
-url = https://github.com/apirrone/mini_BDX_runtime
-description = Runtime code for mini BDX
+url = https://github.com/apirrone/Open_Duck_Mini_Runtime
+description = Runtime code for Open Duck Mini
 long_description = file: README.md
 long_description_content_type = text/markdown
 
@@ -14,7 +14,7 @@ packages = find:
 zip_safe = True
 include_package_data = True
 package_dir=
-    =mini_bdx_runtime
+    =src
 install_requires =
     RPi.GPIO==0.7.1
     rustypot==0.1.0
@@ -29,7 +29,7 @@ install_requires =
     # adafruit_extended_bus
     
 [options.packages.find]
-where=mini_bdx_runtime
+where=src
 
 [options.package_data]
 
@@ -37,3 +37,4 @@ where=mini_bdx_runtime
 
 [options.entry_points]
 console_scripts =
+    walk = open_duck_mini_runtime.walk:main
