+++ src/open_duck_mini_runtime/assets/beep2.wav
