+++ src/open_duck_mini_runtime/assets/happy1.wav
