+++ src/open_duck_mini_runtime/assets/happy2.wav
