+++ src/open_duck_mini_runtime/assets/happy3.wav
