+++ src/open_duck_mini_runtime/assets/lamp3.wav
