+++ src/open_duck_mini_runtime/assets/polynomial_coefficients.pkl
