+++ src/open_duck_mini_runtime/camera.py
