+++ tools/calibrate_imu.py
@@ -1,5 +1,5 @@
 
-from mini_bdx_runtime.raw_imu import Imu
+from open_duck_mini_runtime.raw_imu import Imu
 
 if __name__ == "__main__":
     imu = Imu(50, calibrate=True, upside_down=False)
